Update documentation, reorganice folders for better autodocumentation
